[Tech] create useDrawLayer and use it for dashboard, vigilance areas, recent activity and mission
