+++ frontend/src/features/RecentActivity/components/Layers/DrawRecentActivityLayer.tsx
@@ -1,23 +1,15 @@
-import { drawStyle } from '@features/map/layers/styles/draw.style'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useAppSelector } from '@hooks/useAppSelector'
-import { useDrawVectorLayer } from '@hooks/useDrawVectorLayer'
-import {
-  getOLTypeAndGeometryFunctionFromInteractionType,
-  resetDrawInteractions,
-  resetModifyInteractions
-} from '@utils/drawFunctions'
-import { convertToGeoJSONGeometryObject } from 'domain/entities/layers'
+import { useDrawLayer } from '@hooks/useDrawLayer'
 import { Layers } from 'domain/entities/layers/constants'
 import { drawFeature } from 'domain/use_cases/draw/drawFeature'
-import { isEmpty } from 'lodash'
-import { Modify } from 'ol/interaction'
-import Draw from 'ol/interaction/Draw'
-import React, { useCallback, useEffect } from 'react'
+import { DrawEvent } from 'ol/interaction/Draw'
+import React from 'react'
 
 import { recentActivityActions } from '../../slice'
 
 import type { BaseMapChildrenProps } from '@features/map/BaseMap'
+import type { GeoJSON } from 'domain/types/GeoJSON'
 
 function UnmemoizeDrawRecentActivityLayer({ map }: BaseMapChildrenProps) {
   const dispatch = useAppDispatch()
@@ -25,77 +17,13 @@ function UnmemoizeDrawRecentActivityLayer({ map }: BaseMapChildrenProps) {
   const interactionType = useAppSelector(state => state.recentActivity.interactionType)
   const isDrawing = useAppSelector(state => state.recentActivity.isDrawing)
 
-  const { drawVectorSourceRef, feature, vectorLayerRef, vectorSourceRef } = useDrawVectorLayer(
+  useDrawLayer({
     geometry,
-    Layers.DRAW_RECENT_ACTIVITY.code
-  )
-
-  useEffect(() => {
-    map.getLayers().push(vectorLayerRef.current)
-
-    return () => {
-      if (map) {
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        map.removeLayer(vectorLayerRef.current)
-      }
-    }
-  }, [map, vectorLayerRef])
-
-  const setGeometryOnModifyEnd = useCallback(
-    event => {
-      const nextGeometry = event.features.item(0).getGeometry()
-
-      if (nextGeometry) {
-        const convertedGeometry = convertToGeoJSONGeometryObject(nextGeometry)
-        dispatch(recentActivityActions.setGeometry(convertedGeometry))
-      }
-    },
-    [dispatch]
-  )
-
-  useEffect(() => {
-    if (isEmpty(feature) || !isDrawing) {
-      return undefined
-    }
-
-    resetModifyInteractions(map)
-    vectorSourceRef.current.clear(true)
-    drawVectorSourceRef.current.clear(true)
-    vectorSourceRef.current.addFeature(feature)
-    const modify = new Modify({
-      source: vectorSourceRef.current
-    })
-    map?.addInteraction(modify)
-
-    modify.on('modifyend', setGeometryOnModifyEnd)
-
-    return () => {
-      if (map) {
-        map.removeInteraction(modify)
-        modify.un('modifyend', setGeometryOnModifyEnd)
-      }
-    }
-  }, [map, feature, setGeometryOnModifyEnd, isDrawing, vectorSourceRef, drawVectorSourceRef])
-
-  useEffect(() => {
-    if (!map || !isDrawing) {
-      return undefined
-    }
-
-    resetDrawInteractions(map)
-    const { geometryFunction, geometryType } = getOLTypeAndGeometryFunctionFromInteractionType(interactionType)
-
-    const draw = new Draw({
-      geometryFunction,
-      source: drawVectorSourceRef.current,
-      stopClick: true,
-      style: drawStyle,
-      type: geometryType
-    })
-
-    map.addInteraction(draw)
-
-    draw.on('drawend', event => {
+    interactionType,
+    isDrawing,
+    layerName: Layers.DRAW_RECENT_ACTIVITY.code,
+    map,
+    onDrawEnd: (event: DrawEvent) => {
       dispatch(
         drawFeature(
           event.feature,
@@ -103,20 +31,9 @@ function UnmemoizeDrawRecentActivityLayer({ map }: BaseMapChildrenProps) {
           state => state.recentActivity.drawedGeometry
         )
       )
-      event.stopPropagation()
-      drawVectorSourceRef.current.clear(true)
-    })
-
-    return () => {
-      if (map) {
-        map.removeInteraction(draw)
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        vectorSourceRef.current.clear(true)
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        drawVectorSourceRef.current.clear(true)
-      }
-    }
-  }, [map, dispatch, interactionType, isDrawing, drawVectorSourceRef, vectorSourceRef])
+    },
+    onModifyEnd: (geom: GeoJSON.Geometry) => dispatch(recentActivityActions.setGeometry(geom))
+  })
 
   return null
 }
