+++ frontend/src/features/map/layers/DrawLayer.ts
@@ -1,160 +1,36 @@
-import { useDrawVectorLayer } from '@hooks/useDrawVectorLayer'
-import { isEmpty } from 'lodash'
-import { Modify } from 'ol/interaction'
-import Draw, { createBox, createRegularPolygon, type GeometryFunction } from 'ol/interaction/Draw'
-import React, { useCallback, useEffect } from 'react'
+import { useDrawLayer } from '@hooks/useDrawLayer'
+import { Layers } from 'domain/entities/layers/constants'
+import { DrawEvent } from 'ol/interaction/Draw'
+import React from 'react'
 
-import { drawStyle } from './styles/draw.style'
-import { InteractionType, OLGeometryType } from '../../../domain/entities/map/constants'
 import { addFeatureToDrawedFeature } from '../../../domain/use_cases/draw/addFeatureToDrawedFeature'
 import { setGeometry } from '../../../domain/use_cases/draw/setGeometry'
 import { useAppDispatch } from '../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../hooks/useAppSelector'
 
 import type { BaseMapChildrenProps } from '../BaseMap'
+import type { GeoJSON as GeoJSONType } from 'domain/types/GeoJSON'
 import type Geometry from 'ol/geom/Geometry'
 
 function UnmemoizedDrawLayer({ map }: BaseMapChildrenProps) {
   const dispatch = useAppDispatch()
   const geometry = useAppSelector(state => state.draw.geometry)
   const interactionType = useAppSelector(state => state.draw.interactionType)
-  const listener = useAppSelector(state => state.draw.listener)
 
-  const { drawVectorSourceRef, feature, vectorLayerRef, vectorSourceRef } = useDrawVectorLayer(geometry, 'DRAW')
-
-  useEffect(() => {
-    if (map) {
-      map.getLayers().push(vectorLayerRef.current)
-    }
-
-    return () => {
-      if (map) {
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        map.removeLayer(vectorLayerRef.current)
-      }
-    }
-  }, [map, vectorLayerRef])
-
-  const setGeometryOnModifyEnd = useCallback(
-    event => {
-      const nextGeometry = event.features.item(0).getGeometry()
-      if (nextGeometry) {
-        dispatch(setGeometry(nextGeometry as Geometry))
-      }
-    },
-    [dispatch]
-  )
-
-  useEffect(() => {
-    vectorSourceRef.current.clear(true)
-    drawVectorSourceRef.current.clear(true)
-    if (isEmpty(feature) || !interactionType) {
-      return undefined
-    }
-
-    resetModifyInteractions(map)
-
-    vectorSourceRef.current.addFeature(feature)
-    const modify = new Modify({
-      source: vectorSourceRef.current
-    })
-    map?.addInteraction(modify)
-
-    modify.on('modifyend', setGeometryOnModifyEnd)
-
-    return () => {
-      if (map) {
-        map.removeInteraction(modify)
-        modify.un('modifyend', setGeometryOnModifyEnd)
-      }
-    }
-  }, [map, feature, interactionType, setGeometryOnModifyEnd, vectorSourceRef, drawVectorSourceRef])
-
-  useEffect(() => {
-    if (!map || !interactionType) {
-      return undefined
-    }
-
-    resetDrawInteractions(map)
-    const { geometryFunction, geometryType } = getOLTypeAndGeometryFunctionFromInteractionType(interactionType)
-
-    const draw = new Draw({
-      geometryFunction,
-      source: drawVectorSourceRef.current,
-      stopClick: true,
-      style: drawStyle,
-      type: geometryType
-    })
-
-    map.addInteraction(draw)
-
-    draw.on('drawend', event => {
+  useDrawLayer({
+    geometry,
+    interactionType,
+    isDrawing: !!interactionType,
+    layerName: Layers.DRAW_DASHBOARD.code,
+    map,
+    onDrawEnd: (event: DrawEvent) => {
       dispatch(addFeatureToDrawedFeature(event.feature))
-      event.stopPropagation()
-      drawVectorSourceRef.current.clear(true)
-    })
-
-    return () => {
-      if (map) {
-        map.removeInteraction(draw)
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        vectorSourceRef.current.clear(true)
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        drawVectorSourceRef.current.clear(true)
-      }
-    }
-  }, [map, dispatch, listener, interactionType, drawVectorSourceRef, vectorSourceRef])
+    },
+    onModifyEnd: (geom: GeoJSONType.Geometry) => dispatch(setGeometry(geom as unknown as Geometry)),
+    withConversionToGeoJSONGeometryObject: false
+  })
 
   return null
-}
-
-function getOLTypeAndGeometryFunctionFromInteractionType(interactionType: InteractionType | null): {
-  geometryFunction: GeometryFunction | undefined
-  geometryType: OLGeometryType
-} {
-  switch (interactionType) {
-    case InteractionType.SQUARE:
-      return {
-        geometryFunction: createBox(),
-        geometryType: OLGeometryType.CIRCLE
-      }
-    case InteractionType.CIRCLE:
-      return {
-        geometryFunction: createRegularPolygon(),
-        geometryType: OLGeometryType.CIRCLE
-      }
-    case InteractionType.POLYGON:
-      return {
-        geometryFunction: undefined,
-        geometryType: OLGeometryType.POLYGON
-      }
-    case InteractionType.POINT:
-      return {
-        geometryFunction: undefined,
-        geometryType: OLGeometryType.POINT
-      }
-    default:
-      return {
-        geometryFunction: undefined,
-        geometryType: OLGeometryType.POINT
-      }
-  }
-}
-
-function resetModifyInteractions(map) {
-  map.getInteractions().forEach(interaction => {
-    if (interaction instanceof Modify) {
-      interaction.setActive(false)
-    }
-  })
-}
-
-function resetDrawInteractions(map) {
-  map.getInteractions().forEach(interaction => {
-    if (interaction instanceof Draw) {
-      interaction.setActive(false)
-    }
-  })
 }
 
 export const DrawLayer = React.memo(UnmemoizedDrawLayer)
